Add files via upload
The Student Result Analysis project aims to provide a comprehensive analysis of student performance using data visualization techniques. This project helps educators and stakeholders understand the overall performance of students, identify trends, and make data-driven decisions to enhance the learning experience. The project includes the creation of a heatmap and a pie chart to represent different aspects of student results.